Refactor POST /api/collections/collectionId
